split labeling into main/sherlockwrapper/fileio/misc
